requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-dotenv
google-generativeai
psycopg2-binary